refactor(hub-common): abstract out common internal helper and add TSDoc comments
affects: @esri/hub-common
