+++ packages/common/src/sites/feeds/types.ts
@@ -1,6 +1,17 @@
+/**
+ * Represents the feed format types that are supported in Hub
+ */
 export type FeedFormat = "dcat-us" | "dcat-ap" | "rss";
 
+/**
+ * Configuration object for feeds. Stores feed templates for various feed formats and versions
+ * as well as miscellaneous configuration options. Always interact with this object through the
+ * configuration helpers (getFeedTemplate(), setFeedTemplate(), etc.) when possible.
+ */
 export interface IFeedsConfiguration extends IWithFeedTemplatePaths {
+  /**
+   * Whether or not the feeds capability as a whole has been disabled
+   */
   disabled?: boolean;
 }
 
@@ -13,6 +24,23 @@ type FeedTemplatePath =
   | DcatAPTemplatePaths
   | RssTemplatePaths;
 
-type DcatUSTemplatePaths = "dcatUS1X" | "dcatUS11";
-type DcatAPTemplatePaths = "dcatAP2XX" | "dcatAP201";
-type RssTemplatePaths = "rss2";
+type DcatUSTemplatePaths =
+  // Original path for DCAT-US 1.1 configuration. We have an in-memory migration
+  // that automatically migrates this value to an updated path. Some older sites
+  // that haven't been saved in a while will still have this path present.
+  | "dcatUS11"
+  // Updated path for the primary DCAT-US 1.X template. This should be the default
+  // template for all DCAT-US 1.X feeds (unless overriden by a more specific version)
+  | "dcatUS1X";
+
+type DcatAPTemplatePaths =
+  // Original path for DCAT-AP 2.0.1 configuration. We no longer support editing 2.0.1
+  // templates, but we still expose an API endpoint for the 2.0.1 version. We'll
+  // need to keep this value for backwards compatibility.
+  | "dcatAP201"
+  // Updated path for the primary DCAT-AP 2.X template. This should be the default
+  // template for all DCAT-AP 2.X.X feeds (unless overriden by a more specific version).
+  // 2.1.1 is the latest version of 2.X.X as of this writing.
+  | "dcatAP2XX";
+
+type RssTemplatePaths = "rss2"; // Original path for the RSS 2.0 feed
